Adapt tests to new registry fetching; extract common fixture.
