+++ tests/cli/ursula/test_stake_via_allocation_contract.py
@@ -53,22 +53,8 @@ from nucypher.utilities.sandbox.middleware import MockRestMiddleware
 
 
 @pytest.fixture(autouse=True, scope='module')
-def patch_fetch_latest_publication(test_registry):
-    empty_allocation_escrow_deployer = PreallocationEscrowDeployer(registry=test_registry)
-    allocation_contract_abi = empty_allocation_escrow_deployer.get_contract_abi()
-    allocation_template = {
-        "BENEFICIARY_ADDRESS": ["ALLOCATION_CONTRACT_ADDRESS", allocation_contract_abi]
-    }
-    new_fetch_result = json.dumps(allocation_template).encode()
-
-    original_fetch = IndividualAllocationRegistry.fetch_latest_publication
-
-    def new_fetch(*args, **kwargs):
-        return new_fetch_result
-
-    IndividualAllocationRegistry.fetch_latest_publication = new_fetch
-    yield
-    IndividualAllocationRegistry.fetch_latest_publication = original_fetch
+def patch_individual_allocation_fetch_latest_publication(_patch_individual_allocation_fetch_latest_publication):
+    pass
 
 
 @pytest.fixture(scope='module')
